+++ CHANGELOG.md
@@ -6,12 +6,15 @@
 - Add invalidatie columns (verwijderd/duplicaat + reden) to the eredienst mandatarissen report [OP-3859]
 - VDDS: change config (not retrieving vendor from submission itself, ignore sessions when looking for vendors)
 - bump ipdc-bookmarks [DL-7577]
+- update `ipdc-ldes-consumer` to `feature-query-delay` tag [DL-7589]
+- `ipdc-ldes-consumer`: increase SPARQL_BATCH_SIZE to `1000`
 
 ## Deploy notes
 
 ```
 drc pull delta-producer-publication-graph-maintainer && drc up -d delta-producer-publication-graph-maintainer
 drc pull ipdc-bookmarks && drc up -d ipdc-bookmarks
+drc pull ipdc-ldes-consumer && drc up -d ipdc-ldes-consumer
 drc exec delta-producer-background-jobs-initiator curl -X POST http://localhost/worship-services-sensitive/healing-jobs
 drc exec vendor-data-distribution curl -X POST http://localhost/heal
 drc restart report-generation